requirements.txt
openai>=1.6.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pydantic>=2.0.0
colorama>=0.4.6
python-dotenv>=1.0.0
tenacity>=8.2.0
